// src/module/schedule-management/time-requests/utils/enums.ts
export enum tabParams {
	EXTENDED_TIME = "extended-time",
	MIDDAY_STOP = "midday-stop",
	FINGERPRINT_APPROVAL = "fingerprint-approval",
}

export enum requestStatusTypes {
	APPROVED = "approved",
	REJECTED = "rejected",
	PENDING = "pending",
}

export enum fingerprintApprovalStatus {
	PENDING = "PENDING",
	ACCEPTED = "ACCEPTED",
	DECLINED = "DECLINED",
}

export enum searchParamsTypes {
	TAB = "tab",
	SEARCH = "search",
	STATUS = "status",
	TYPE = "type",
	DATE = "date",
}